Se agrego hook ventala modal y se agrgaron los mensajes de confirmación para eliminar usuario y cerrar sesión

+++ src/components/ItemTable/ItemTable.jsx
@@ -1,18 +1,53 @@
+/* eslint-disable no-restricted-globals */
 /* eslint-disable react/prop-types */
 import React from 'react';
+import Swal from 'sweetalert2';
+import withReactContent from 'sweetalert2-react-content';
 import { deleteUserRequest } from '../../API/requestHTTP';
 import './ItemTable.scss';
 
-function ItemTable({ username, userRole, id, setValueForm, setEdit }) {
+function ItemTable({
+  username,
+  userRole,
+  id,
+  setValueForm,
+  setEdit,
+  openModal,
+}) {
+  const MySwal = withReactContent(Swal);
+
   const onDelete = () => {
     const token = localStorage.getItem('token');
-    deleteUserRequest(token, id)
-      .then((response) => console.log(response))
-      .catch((error) => console.log(error));
+    MySwal.fire({
+      title: '¿Estás seguro que deseas eliminar el usuario?',
+      text: 'No podrás deshacer esto!',
+      icon: 'warning',
+      showCancelButton: true,
+      confirmButtonColor: '#ff7272',
+      cancelButtonColor: '#8c8787',
+      confirmButtonText: 'Si, elimínalo!',
+      customClass: {
+        popup: 'delete-user-alert',
+      },
+    }).then((result) => {
+      if (result.isConfirmed) {
+        deleteUserRequest(token, id)
+          .then((response) => console.log(response))
+          .catch((error) => console.log(error));
+        Swal.fire({
+          title: 'Usuario eliminado!',
+          icon: 'success',
+          confirmButtonColor: '#fdad4e',
+        });
+      }
+    });
   };
 
   // On Edit Function
   const onEdit = () => {
+    if (screen.width < 1000) {
+      openModal();
+    }
     setEdit(true);
     setValueForm({
       email: username,
